refactor: extract config parsing and validation
Rollback trigger: None
